Fixed return type from function distribution_total_counter from double to uint64_t

+++ src/daemon/distribution.c
@@ -317,7 +317,7 @@ double distribution_total_sum(distribution_t *dist) {
   return dist->total_sum; // should I add mutex here?
 }
 
-double distribution_total_counter(distribution_t *dist) {
+uint64_t distribution_total_counter(distribution_t *dist) {
   if (dist == NULL) {
     return NAN;
   }
